refactor: Categories Reducer & Selector
