+++ src/index.js
@@ -5,8 +5,6 @@ import { Provider } from 'react-redux';
 import { store } from './store/store';
 
 import App from './App';
-import { UserProvider } from './contexts/user/user.context'
-import { CategoriesProvider } from './contexts/categories/categories.context';
 import { CartProvider } from './contexts/cart/cart.context';
 
 import './index.scss';
@@ -17,13 +15,9 @@ root.render(
   <React.StrictMode>
     <Provider store = {store}>
       <BrowserRouter>
-        <UserProvider>
-          <CategoriesProvider>
             <CartProvider>
               <App />
             </CartProvider>
-          </CategoriesProvider>
-        </UserProvider>
       </BrowserRouter>
     </Provider>
   </React.StrictMode>
