+++ src/routes/shop/shop.component.jsx
@@ -1,9 +1,26 @@
+import { useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
+
+import { useDispatch } from 'react-redux';
 
 import CategoriesPreview from '../categories-preview/categories-preview.component';
 import Category from '../category/category.component';
+import {
+  getCategoriesAndDocuments
+} from '../../utils/firebase/firebase.utils';
+import { setCategoriesMap } from '../../store/categories/category.action';
 
 const Shop = () => {
+  const dispatch = useDispatch();
+  useEffect(() => {
+    const getCategoriesMap = async () => {
+      const categoryMap = await getCategoriesAndDocuments('categories');
+      dispatch(setCategoriesMap(categoryMap));
+    };
+
+    getCategoriesMap();
+  }, []);
+
   return (
     <Routes>
       <Route index element={<CategoriesPreview />} />
