+++ src/store/categories/category.action.js
@@ -1,0 +1,4 @@
+import { createAction } from '../../utils/reducer/reducer.utils';
+import { CATEGORIES_ACTION_TYPES } from './category.types';
+
+export const setCategoriesMap = (categoriesMap) => createAction(CATEGORIES_ACTION_TYPES.SET_CATEGORIES_MAP, categoriesMap);
