+++ src/store/root-reducer.js
@@ -1,7 +1,9 @@
 import { combineReducers } from "redux";
 
 import { userReducer } from "./user/user.reducer";
+import { categoriesReducer } from "./categories/category.reducer";
 
 export const rootReducer = combineReducers({
     user: userReducer,
+    categories: categoriesReducer,
     });
